requirements.txt
# requirements.txt
# --- Backend Core ---
fastapi
uvicorn[standard]
python-multipart
sqlalchemy
pymysql
cryptography
python-dotenv

# --- AI Services ---
faster-whisper
# transformers  <-- (EXAONE GGUF 구동 시엔 필수는 아니지만, 다른 서비스 위해 유지)
transformers
torch
sentencepiece
accelerate
pydub

# [New] LG EXAONE GGUF 구동용
llama-cpp-python
huggingface-hub

# --- Frontend & Visualization ---
streamlit
requests
pandas
altair

# --- Monitoring ---
prometheus-fastapi-instrumentator
